Implement Gallery `Feed` page
A gallery feed page is just actually the news feed of a social platform.

It is essential because it provides users with a centralized, continuously
updated stream of content from the community.

This feature enhances user engagement by keeping them informed and connected
with the latest happenings, fostering interaction and community involvement.